Revert "Reafactoring sharif"
