+++ backend/api/chat.api.js
@@ -1,11 +1,5 @@
 import OpenAI from "openai";
 
-/**
- * System prompt for resume parsing
- * Provides detailed instructions to the AI for extracting structured data from resumes
- * Defines output format and processing rules
- * @type {String}
- */
 const systemPrompt = `
 
 You are an intelligent resume parsing assistant. Your task is to extract and structure information from unstructured resume text. The text may be disorganized, with related elements appearing in different locations. Your goal is to correctly associate and organize all relevant pieces of information while maintaining accuracy.
@@ -63,38 +57,24 @@ Your response should ONLY contain the formatted JSON data—no explanations, not
 
 `;
 
-/**
- * Chat API endpoint handler
- * Processes resume text using AI to extract structured information
- * Uses OpenRouter API to access LLaMA 3.2 model
- * @param {Object} req - Express request object containing resume text
- * @param {Object} res - Express response object
- * @returns {Object} JSON response with parsed resume data
- */
 export default async function chat(req, res) {
-  const openai = new OpenAI({
-      baseURL: "https://openrouter.ai/api/v1",
-      apiKey: process.env.OPENROUTER_API_KEY,
-  })
-  const { query } = req.body;
-  try {
-      /**
-       * Send resume text to AI model for parsing
-       */
-      const completion = await openai.chat.completions.create({
-          model: "meta-llama/llama-3.2-3b-instruct:free",
-          messages: [
-              { role: "system", content: systemPrompt },
-              { role: "user", content: query },
-          ],
-      })
-      return res.status(200).json({ success: true, data: completion });
-  } catch (err) {
-      /**
-       * Handle and log API errors
-       */
-      console.log(err);
-      res.status(500).json({ success: false, message: "Internal server error "+err });
-  }
-  
+    const openai = new OpenAI({
+        baseURL: "https://openrouter.ai/api/v1",
+        apiKey: process.env.OPENROUTER_API_KEY,
+    })
+    const { query } = req.body;
+    try {
+        const completion = await openai.chat.completions.create({
+            model: "meta-llama/llama-3.2-3b-instruct:free",
+            messages: [
+                { role: "system", content: systemPrompt },
+                { role: "user", content: query },
+            ],
+        })
+        return res.status(200).json({ success: true, data: completion });
+    } catch (err) {
+        console.log(err);
+        res.status(500).json({ success: false, message: "Internal server error "+err });
+    }
+    
 }
