+++ backend/controllers/application.controller.js
@@ -2,8 +2,8 @@ import mongoose, { get } from "mongoose";
 import Application from "../models/application.model.js";
 import Job from "../models/job.model.js";
 import User from "../models/user/user.model.js";
-import codeAssessment from "../models/codeAssessment.model.js";
-import codeSubmission from "../models/codeSubmission.model.js";
+import codeAssessment from "../models/codeAssessment.js";
+import codeSubmission from "../models/codeSubmission.js";
 
 const createResponse = (success, message, data = null) => ({
     success,
@@ -21,7 +21,20 @@ const handleError = (res, error, defaultMessage = "Server error") => {
 
 export const applicationController = {
   
-    
+    async getApplication(req, res) {
+        try {
+            const { jobId } = req.params;
+            const application = await Application.findOne({ job: jobId, applicant: req.uid });
+
+            if (!application) {
+                return res.status(404).json(createResponse(false, "Application not found"));
+            }
+
+            return res.status(200).json(createResponse(true, "Application retrieved successfully", application));
+        } catch (error) {
+            return handleError(res, error, "Error retrieving application");
+        }
+    },
 
     async updateApplicationStatus(req, res) {
         try {
