+++ backend/fixtures/jobQuestionFixtures.js
@@ -1,105 +1,40 @@
-/**
- * Job application question fixtures
- * Contains common interview questions and sample responses
- * Used for generating application forms and screening questions
- * @type {Array<Object>}
- */
 export const jobQuestions = [
-    /**
-     * Company selection rationale question
-     * Assesses candidate's research and alignment with company values
-     * @type {Object}
-     */
     {
         questionText: "Why did you choose this company?",
         sampleAnswer: "I chose this company because of its innovative approach to technology and its commitment to fostering a collaborative work environment. I admire how the company values continuous learning and growth, which aligns with my personal and professional goals."
     },
-    
-    /**
-     * Position motivation question
-     * Evaluates candidate's interest and fit for specific role
-     * @type {Object}
-     */
     {
         questionText: "What motivates you to apply for this position?",
         sampleAnswer: "I am motivated to apply for this position because it aligns perfectly with my skills and career aspirations. I am excited about the opportunity to contribute to impactful projects and grow within a dynamic team."
     },
-    
-    /**
-     * Strengths assessment question
-     * Identifies candidate's self-awareness and relevant abilities
-     * @type {Object}
-     */
     {
         questionText: "What are your strengths that make you a good fit for this role?",
         sampleAnswer: "My strengths include strong problem-solving skills, effective communication, and the ability to work well in a team. I also have a proven track record of meeting deadlines and delivering high-quality work."
     },
-    
-    /**
-     * Challenge handling question
-     * Evaluates resilience and problem-solving approach
-     * @type {Object}
-     */
     {
         questionText: "How do you handle challenges or setbacks at work?",
         sampleAnswer: "I handle challenges by staying calm and focused, analyzing the situation, and developing a plan to address the issue. I believe that setbacks are opportunities to learn and improve, and I always strive to find constructive solutions."
     },
-    
-    /**
-     * Career goals question
-     * Assesses long-term aspirations and commitment
-     * @type {Object}
-     */
     {
         questionText: "Where do you see yourself in five years?",
         sampleAnswer: "In five years, I see myself growing within the company, taking on more responsibilities, and contributing to larger projects. I aim to develop my expertise further and potentially move into a leadership role."
     },
-    
-    /**
-     * Company culture awareness question
-     * Evaluates research and cultural fit
-     * @type {Object}
-     */
     {
         questionText: "What do you know about our company culture?",
         sampleAnswer: "I understand that your company culture emphasizes collaboration, innovation, and employee development. I appreciate how the company encourages creativity and supports work-life balance, which I believe fosters a productive and positive work environment."
     },
-    
-    /**
-     * Time management question
-     * Assesses organizational skills and efficiency
-     * @type {Object}
-     */
     {
         questionText: "How do you prioritize your tasks and manage your time?",
         sampleAnswer: "I prioritize tasks by assessing their urgency and importance, and I use tools like to-do lists and project management software to stay organized. I also allocate specific time blocks for focused work and regularly review my progress to ensure I meet deadlines."
     },
-    
-    /**
-     * Achievement question
-     * Identifies significant accomplishments and success patterns
-     * @type {Object}
-     */
     {
         questionText: "What is your greatest professional achievement?",
         sampleAnswer: "My greatest professional achievement was leading a team to successfully complete a complex project ahead of schedule. This experience taught me valuable lessons in leadership, teamwork, and project management."
     },
-    
-    /**
-     * Industry knowledge question
-     * Evaluates commitment to professional growth
-     * @type {Object}
-     */
     {
         questionText: "How do you stay updated with industry trends?",
         sampleAnswer: "I stay updated by following industry blogs, attending webinars, and participating in professional development courses. I also enjoy networking with peers to exchange knowledge and insights."
     },
-    
-    /**
-     * Value proposition question
-     * Assesses candidate's confidence and unique selling points
-     * @type {Object}
-     */
     {
         questionText: "Why should we hire you for this position?",
         sampleAnswer: "You should hire me because I bring a unique combination of skills, experience, and enthusiasm that aligns with the requirements of this role. I am committed to delivering high-quality work and contributing to the success of the team and the company."
