+++ backend/fixtures/locationFixtures.js
@@ -1,9 +1,3 @@
-/**
- * Location fixture data
- * Contains a list of global cities for job locations
- * Used for generating job postings and candidate profiles
- * @type {Array<String>}
- */
 export const locations = [
     'London',
     'Edinburgh',
