+++ frontend/vite.config.js
@@ -6,7 +6,7 @@ export default defineConfig({
   plugins: [react()],
   server:{
     proxy: {
-      '/api': 'http://localhost:5001',
+      '/api': 'http://localhost:5000',
     }
   },
   define: {
